LayoutTests/imported/w3c/web-platform-tests/uievents/order-of-events/mouse-events/wheel-basic-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unknown source type "wheel".
Description: Verifies that wheel events are captured and sent

Instructions:

Place your mouse pointer over the top box
Scroll the mouse wheel to change the box color
Move the mouse pointer to the second box
Scroll the mouse wheel again to change this box's color
Test Passes if the box turns green and the word 'PASS' appears below


Harness Error (FAIL), message = Unhandled rejection: Unknown source type "wheel".

NOTRUN wheel event test
